requirements.txt
torch==2.4.1+cpu
torchvision==0.19.1+cpu
transformers>=4.48.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
streamlit==1.38.0
pymupdf>=1.24.10
beautifulsoup4>=4.12.3
pandas>=2.2.2
numpy==1.26.4
pydantic>=2.9.2
python-dotenv>=1.0.1
datasets>=2.21.0
sec-edgar-downloader>=5.1.0
sec-parser==0.58.1
lxml>=5.3.0
tqdm>=4.66.5
uvicorn>=0.30.6
fastapi>=0.115.0
python-dotenv
